modules/human-paddle-control/requirements.txt
imageio==2.31.1
matplotlib==3.7.1
numpy==1.25.0
opencv-python-headless==4.8.1.78
paho-mqtt==1.6.1
pyrealsense2==2.54.1.5217
transitions==0.9.2
jsonschema==4.23.0
